add age

+++ src/views/Treatment.vue
@@ -2,10 +2,14 @@
   <div
     @mousemove.once="
       getNewTreatmentData();
-      getDayTreatment();
+      getDayTreatment(dayData);
     "
     class="mt-2"
   >
+    <div class="text-center">
+      <p>年齢: {{ ageData }}</p>
+    </div>
+    <hr />
     <h4>処置一覧</h4>
     <div v-if="defaultBoolean">
       <button @click="inversionBoolean" class="btn btn-warning m-0 py-0 px-1">
@@ -42,11 +46,19 @@
         </div>
 
         <label class="col-form-label col-5">選択内容更新:</label>
-        <div class="col-10 col-lg-6 mb-3">
-          <textarea
+        <div class="col-10 col-lg-6 mb-2">
+          <input
+            type="text"
+            autocomplete="on"
+            list="record"
             v-model="updateTreatmentText"
             class="form-control"
-          ></textarea>
+          />
+          <datalist id="record">
+            <option v-for="(newTreatment, key) in newTreatmentObj" :key="key">
+              {{ newTreatment.value.treatment }}
+            </option>
+          </datalist>
         </div>
 
         <button
@@ -59,7 +71,7 @@
           @click="deleteTreatmentData(updateTreatmentID)"
           class="btn btn-primary mx-2"
         >
-          削除
+          選択した処置記録を削除
         </button>
       </div>
     </div>
@@ -73,11 +85,10 @@
     <hr />
 
     <div>
-      <label class="col-2 col-form-label">日付: </label>
-      <div class="col-8 col-lg-3">
+      <label class="col-4 col-form-label">処置記録登録: </label>
+      <div class="col-8 col-lg-3 mb-3">
         <input type="datetime-local" v-model="day" class="form-control" />
       </div>
-      <label class="col-2 col-form-label">処置選択</label>
       <div class="col-10 col-lg-3 mb-3">
         <select v-model="dayTreatment" class="form-select form-select-sm">
           <option value="">選択してください</option>
@@ -93,6 +104,36 @@
       <button @click="addDayTreatmentData" class="btn btn-primary">登録</button>
     </div>
 
+    <hr />
+
+    <div>
+      <label lass="col-2 col-form-label">各月の処置記録抽出:</label>
+      <div class="col-6 my-2">
+        <input type="month" v-model="selectDayValue" class="form-control" />
+      </div>
+      <button
+        @click="getDayTreatment(selectDayValue)"
+        class="btn btn-primary px-1"
+      >
+        {{ selectDayValue }}月分表示
+      </button>
+    </div>
+
+    <hr />
+
+    <div>
+      <label class="col-6 col-form-label">日付指定処置記録抽出:</label>
+
+      <div class="col-6 mb-2">
+        <input type="date" v-model="dayKeywordFirst" class="form-control" />
+        <p class="m-0 p-0">から</p>
+        <input type="date" v-model="dayKeywordSecond" class="form-control" />
+      </div>
+      <button class="btn btn-primary px-1" @click="getMonthTreatment">
+        {{ dayKeywordFirst }}-{{ dayKeywordSecond }}分表示
+      </button>
+      <button class="btn btn-primary px-1 mx-2" @click="creaDay">クリア</button>
+    </div>
     <hr />
 
     <label class="col-6 col-form-label">キーワード検索: </label>
@@ -164,6 +205,7 @@ import { Mixins, Prop } from "vue-property-decorator";
 export default class treatment extends Mixins(MixinLogger) {
   @Prop() id!: number;
   @Prop() userID!: string;
+  @Prop() birthday!: string;
 
   newTreatment = ""; //新規処置入力用の値
   newTreatmentObj = {}; //新規処置入力で登録した値を格納するオブジェクト
@@ -206,7 +248,14 @@ export default class treatment extends Mixins(MixinLogger) {
     return dayTreatments;
   }
 
+  created() {
+    this.getAge(this.birthday);
+  }
+
   addNewTreatmentData() {
+    if (this.newTreatment === "") {
+      return alert("新規処置記録を入力してください");
+    }
     //新規処置登録用のメソッド
     this.treatmentsDb
       .doc(String(this.$_uid))
@@ -215,6 +264,7 @@ export default class treatment extends Mixins(MixinLogger) {
         treatmentID: this.$_uid,
       })
       .then(() => {
+        alert("登録しました");
         this.newTreatment = "";
         this.uidCreate();
       });
@@ -238,6 +288,9 @@ export default class treatment extends Mixins(MixinLogger) {
   }
 
   addDayTreatmentData() {
+    if (this.dayTreatment === "" || this.day === "") {
+      return alert("日付、処置を入力してください");
+    }
     this.userTreatmentsDb
       .doc(String(this.$_uid))
       .set({
@@ -247,38 +300,84 @@ export default class treatment extends Mixins(MixinLogger) {
         treatmentID: this.$_uid,
       })
       .then(() => {
+        alert("登録しました");
         this.dayTreatment = "";
         this.uidCreate();
       });
   }
 
   deleteDayTreatmentData(treatmentID: string) {
-    this.userTreatmentsDb.doc(String(treatmentID)).delete();
-    alert("削除しました");
+    this.userTreatmentsDb
+      .doc(String(treatmentID))
+      .delete()
+      .then(() => alert("削除しました"));
   }
 
-  getDayTreatment() {
-    this.userTreatmentsDb.onSnapshot(
-      (querySnapshot) => {
-        const obj: {
-          [key: string]: { value: firebase.firestore.DocumentData };
-        } = {};
-        querySnapshot.forEach((doc) => {
-          obj[doc.id] = { value: doc.data() };
-        });
-        this.dayTreatmentObj = obj;
-      },
-      (error) => {
-        console.log(error.message);
-      }
-    );
+  getDayTreatment(dayValue: string) {
+    if (dayValue === "") {
+      return alert("期間を入力してください");
+    }
+    const startDay = dayValue + "-01";
+    const endDay = dayValue + "-32";
+    this.userTreatmentsDb
+      .where("day", ">=", startDay)
+      .where("day", "<=", endDay)
+      .limit(150)
+      .onSnapshot(
+        (querySnapshot) => {
+          const obj: {
+            [key: string]: { value: firebase.firestore.DocumentData };
+          } = {};
+          querySnapshot.forEach((doc) => {
+            obj[doc.id] = { value: doc.data() };
+          });
+          this.dayTreatmentObj = obj;
+        },
+        (error) => {
+          console.log(error.message);
+        }
+      );
+  }
+
+  getMonthTreatment() {
+    if (this.dayKeywordFirst === "" || this.dayKeywordSecond === "") {
+      return alert("期間を入力してください");
+    }
+    this.userTreatmentsDb
+      .where("day", ">=", this.dayKeywordFirst)
+      .where("day", "<=", this.dayKeywordSecond)
+      .limit(150)
+      .onSnapshot(
+        (querySnapshot) => {
+          const obj: {
+            [key: string]: { value: firebase.firestore.DocumentData };
+          } = {};
+          querySnapshot.forEach((doc) => {
+            obj[doc.id] = { value: doc.data() };
+          });
+          this.dayTreatmentObj = obj;
+        },
+        (error) => {
+          console.log(error.message);
+        }
+      );
+  }
+
+  creaDay() {
+    this.dayKeywordFirst = "";
+    this.dayKeywordSecond = "";
+    this.getDayTreatment(this.dayData);
   }
 
   deleteTreatmentData(uid: string) {
+    if (this.updateTreatmentID === "") {
+      return alert("削除する処置記録を選択してください");
+    }
     this.treatmentsDb
       .doc(String(uid))
       .delete()
       .then(() => {
+        alert("削除しました");
         this.updateTreatmentID = "";
         this.updateTreatmentText = "";
         this.uidCreate();
@@ -286,6 +385,9 @@ export default class treatment extends Mixins(MixinLogger) {
   }
 
   updateTreatmentData(uid: string) {
+    if (this.updateTreatmentID === "" || this.updateTreatmentText === "") {
+      return alert("変更する処置記録の選択、変更した記録を入力してください");
+    }
     this.treatmentsDb
       .doc(String(uid))
       .update({
@@ -293,6 +395,7 @@ export default class treatment extends Mixins(MixinLogger) {
         staffName: this.displayStaffName,
       })
       .then(() => {
+        alert("更新しました");
         this.updateTreatmentID = "";
         this.updateTreatmentText = "";
         this.uidCreate();
